Fix race condition(?) for certbot test
Happens when there are multiple names
being requested and there is a delay
in _handle_challenge_finalize().

For some reason, the order's status is not
being set to READY despite all authorizations
being valid and the request to
finalize fails with a 403.

Fixed by doing another order.finalize() call
in _finalize_order for now.