Modify CA/Broker/Proxy inheritance structure
All three now inherit from a new base class:
AcmeServerBase.

Also fix a bug where the broker was adding
its own certificate (which has been removed now)
to the full pem chain that is delivered
to the client. Add a new column "full_chain"
to the Certificate model in order to easily store it.