Client: fix DNS challenge completion; SSL flag->Boulder; Nonce refetching
The issue with advancing the challenge status to PROCESSING
was the following:
Pebble/Boulder differentiate between a POST-as-GET and a
normal signed POST request. So to tell LE that we want
it to check the TXT record, our payload needs to be an empty
JSON object b"{}" instead of an empty string b"".

closes #17