Server: Use X-Forwarded-Proto header to alter URLs
Our reverse proxy nginx adds the header, so we
check for it and use its value to ensure that
the client gets the URLs prepended with the correct URI scheme.

Otherwise, verification of the JWS header would fail or
the server would return plain http-URLs even though
we may not have enabled it.