Order db model: make proxied_url non-unique
This is to fix LE re-using order URLs, i.e.
returning an old order that's still pending if it has the
same identifiers as the one we're trying to create.

This behavior leads to unique constraint violations.

Should look into whether merging the LE order with our existing one
makes sense, instead of allowing proxied_url to be duplicated.

relevant issue: #28