Docker deployment: Use named volume for resty-auto-ssl data
Anonymous volumes are persisted on
docker-compose down, but they are not
mounted on a subsequent docker-compose up.

https://docs.docker.com/compose/reference/down/

"Anonymous volumes are not removed by default.
However, as they don’t have a stable name,
they will not be automatically mounted by a
subsequent up. For data that needs to persist
between updates, use host or named volumes."